requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
dnspython==2.1.0
Flask==2.0.2
gTTS==2.2.3
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.3
pycparser==2.20
pymongo==3.12.1
PyNaCl==1.4.0
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.0
